fix display of nfts and  refacto nftgrid
