+++ svelte/components/Nft/NftGrid.svelte
@@ -1,18 +1,20 @@
 <script lang="ts">
-  import type { Readable } from "svelte/store";
   import type { NftType } from "lib/ktypes";
 
   import { nftName, nftOpenSeaUrl, addressSame } from "lib/kconfig";
 
   import { nftGetImageLink } from "lib/knft-get-metadata";
 
+  /////////////////////////////////////////////////
+  //  <NftGrid {chainId} {tokenID} {account} {nft} {platform}? />
+  // Display NFT card for Grid mode
+  /////////////////////////////////////////////////
   export let chainId: number;
   export let tokenID: string = "";
   export let account: string = undefined;
 
-  //   export let mainContentDisplayComponent: string;
-
-  export let nfts: Readable<Map<string, NftType>>;
+  export let nft: NftType;
+  export let index;
 
   export let platform: string = "dapp";
   ///////////////////////////////////////////////////////////////////////////////
@@ -31,10 +33,10 @@
   const divMediaVideo = (src: string, small = true) => {
     let video: string;
     if (small) {
-      video = "<video preload=\"metadata\" style=\"border-radius: initial;\">";
+      video = '<video preload="metadata" style="border-radius: initial;">';
     } else {
       video =
-        "<video autoplay=\"true\"  controls=\"\" controlslist=\"nodownload\" loop=\"\" playsinline=\"\" preload=\"metadata\" style=\"border-radius: initial;\">";
+        '<video autoplay="true"  controls="" controlslist="nodownload" loop="" playsinline="" preload="metadata" style="border-radius: initial;">';
     }
     video += `<source src="${src}" type="video/mp4"></video>`;
     return video;
@@ -57,7 +59,7 @@
     } else if (mediaType == "image") {
       div += divMediaImage(mediaSrc);
     } else {
-      div += "<div class=\"media-text\"></div>";
+      div += '<div class="media-text"></div>';
     }
     div += "</a></div>";
 
@@ -66,25 +68,21 @@
   };
 </script>
 
-<div class="row grid-krd">
-  {#each [...$nfts.values()] as nft, index}
-    <div class="col col-xs-12 col-sm-6 col-md-4 col-lg-3">
-      <div class="grid-card-krd" on:click={(evt) => displayNftSolo(evt, nft.tokenID)}>
-        {@html divMedia(nft, index, true)}
+<div class="col col-xs-12 col-sm-6 col-md-4 col-lg-3">
+  <div class="grid-card-krd" on:click={(evt) => displayNftSolo(evt, nft.tokenID)}>
+    {@html divMedia(nft, index, true)}
 
-        <div class="caption">
-          <h3>{nftName(nft)}</h3>
-          {#if addressSame(nft.owner, account)}
-            <a href={nftOpenSeaUrl(chainId, nft)} class="btn btn-small btn-sell" title="Sell" target="_blank"> Sell </a>
-          {:else}
-            <a href={nftOpenSeaUrl(chainId, nft)} class="btn btn-small btn-buy" title="Buy" target="_blank"> Buy </a>
-          {/if}
-        </div>
-      </div>
-
-      {#if platform === "wordpress"}
-        <div>WordPress</div>
+    <div class="caption">
+      <h3>{nftName(nft)}</h3>
+      {#if addressSame(nft.owner, account)}
+        <a href={nftOpenSeaUrl(chainId, nft)} class="btn btn-small btn-sell" title="Sell" target="_blank"> Sell </a>
+      {:else}
+        <a href={nftOpenSeaUrl(chainId, nft)} class="btn btn-small btn-buy" title="Buy" target="_blank"> Buy </a>
       {/if}
     </div>
-  {/each}
+  </div>
+
+  {#if platform === "wordpress"}
+    <div>WordPress</div>
+  {/if}
 </div>
